Extract a json_dumps() function into simpleflow utilities
