+++ simpleflow/swf/stats/pretty.py
@@ -1,5 +1,4 @@
 import sys
-import types
 import operator
 from functools import partial, wraps
 from datetime import datetime
@@ -8,6 +7,7 @@ from tabulate import tabulate
 
 from . import WorkflowStats
 from simpleflow.history import History
+from simpleflow.utils import json_dumps
 
 TEMPLATE = '''
 Workflow Execution {workflow_id}
@@ -56,29 +56,13 @@ def human(values, headers):
     )
 
 
-def _serialize_complex_object(obj):
-    if isinstance(obj, datetime):
-        return obj.isoformat()
-    if isinstance(obj, types.GeneratorType):
-        return [i for i in obj]
-    raise TypeError(
-        "Type %s couldn't be serialized. This is a bug in simpleflow,"
-        " please file a new issue on GitHub!" % type(obj))
-
-
 def jsonify(values, headers):
-    import json
-
     if headers:
-        return json.dumps(
-            [dict(zip(headers, value)) for value in values],
-            default=_serialize_complex_object,
+        return json_dumps(
+            [dict(zip(headers, value)) for value in values]
         )
     else:
-        return json.dumps(
-            values,
-            default=_serialize_complex_object,
-        )
+        return json_dumps(values)
 
 
 DEFAULT_FORMAT = partial(tabular, tablefmt='plain', floatfmt='.2f')
