+++ simpleflow/utils/__init__.py
@@ -1,4 +1,5 @@
 from . import retry  # NOQA
+from .json_dumps import json_dumps
 
 
 def issubclass_(arg1, arg2):
